feat(inventory): 新增库存分页查询功能及相关服务实现
- 添加库存分页查询接口及其实现逻辑- 实现根据产品ID批量查询入库和出库订单项
- 增加通过订单ID列表查询入库和出库订单的功能
- 引入库存数量、出货数量及仓库维度的详细统计
- 支持按产品分类、仓库、产品名称等条件筛选
- 实现库存记录的动态排序与分页返回- 新增库存列表响应结构，包含产品、仓库、单位等信息
- 添加BigDecimal类型支持以提高数值计算精度
- 修复部分字段类型转换问题，如outUnitPerNum和weightPerUnit- 优化库存Facade层数据聚合逻辑，提升查询性能

+++ my-project-backend/src/main/java/com/example/entity/cangku/dto/OutboundOrderItem.java
@@ -6,6 +6,8 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import com.example.entity.dto.BaseModel;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 //出库单明细表
 @Data
 @TableName("ck_outbound_order_item")
@@ -19,7 +21,7 @@ public class OutboundOrderItem extends BaseModel {
     //产品ID
     private Long productId;
     //计划数量
-    private Integer quantity;
+    private BigDecimal quantity;
     //实际数量
     private Integer actualQuantity;
 }
